refactor: restructure codebase and improve analytics modal
- Separate frontend and backend code
- Rename and move `modal.templ` to `analytics_modal.templ` in `cmd/web/components/`
- Rename and move `handlers.go` to `shortener_handlers.go` in `cms/web/handlers/`
- Remove redundancies, especially in the analytics modal
- Apply best practices and test all changes
