+++ cmd/web/components/analytics_modal.templ
@@ -1,11 +1,12 @@
-package shortener
+package components
 
 import (
 	"fmt"
 	"time"
+	"volaticus-go/internal/shortener"
 )
 
-templ RenderAnalyticsModal(analytics *URLAnalytics) {
+templ AnalyticsModal(analytics *shortener.URLAnalytics) {
 	<div class="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
 		<div class="bg-gray-800 rounded-lg p-6 w-full max-w-4xl">
 			<div class="flex justify-between items-center mb-6">
